+++ ovos_utils/skills/templates/active.py
@@ -14,11 +14,3 @@ class ActiveSkill(MycroftSkill):
         self.make_active()
 
 
-class PassiveSkill(ActiveSkill):
-
-    def handle_utterance(self, utterances, lang="en-us"):
-        """ Listen to all utterances passively, eg, take metrics """
-
-    def converse(self, utterances, lang="en-us"):
-        self.handle_utterance(utterances, lang)
-        return False
